fix 修改KNum.NumSign结果类型为int8

+++ number_test.go
@@ -535,7 +535,7 @@ func BenchmarkPow(b *testing.B) {
 func TestLog(t *testing.T) {
 	res1 := KNum.Log(100, 10)
 	res2 := KNum.Log(16, 2)
-	if int(res1) != 1 || int(res2)!=4 {
+	if int(res1) != 1 || int(res2) != 4 {
 		t.Error("Log fail")
 		return
 	}
@@ -600,7 +600,7 @@ func BenchmarkIsEven(b *testing.B) {
 func TestSign(t *testing.T) {
 	var tests = []struct {
 		param    float64
-		expected float64
+		expected int8
 	}{
 		{0, 0},
 		{-1, -1},
